requirements.txt
Django==4.2.7
django-jazzmin==2.6.0
Pillow==11.1.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-storages==1.14.2
psycopg2-binary==2.9.9
django-filter==23.3
django-import-export==3.3.1
django-cleanup==8.0.0
django-widget-tweaks==1.5.0
django-debug-toolbar==4.2.0
boto3==1.34.14
pandas==2.1.4
mimetypes==0.1.0
uuid==1.30